requirements.txt
Flask
GitPython
mutagen
pypandoc
python-dotenv
python-dateutil
requests
twitch-python
apscheduler
toml
python-magic
